refactor(ontology-ui): split workbench monolith into nine region components (M2)
Zero-behavior extraction per goals/ontology-workbench-migration M2:

- Session.workbench.tsx (1149 lines) becomes a 55-line composition of
  nine exported region components (explorer, document, graph, source,
  inspector, sparql, validation, changelog, metrics) rendering the
  same DOM and reading the same atoms
- the five Add-Triple form atoms relocate unchanged from the UI file
  into @beep/ontology-client Session.atoms.ts (exported + documented)
- 3D renderer switch moves into the Graph region's chrome per SPEC
  decision 2; the chromium renderer-toggle screenshot test passes
  unchanged
- zero test-file edits: ontology-ui 2/2, ontology-client 14/14 and the
  browser suite all pass against the extracted regions
- StrictMode finding recorded: MUI X useDisposable reads the fiber's
  StrictEffects bit, so no mount wrapper can fix the invariant; nothing
  applied, residual carries to M3
- rides: documented OSV filter for newly published GHSA-xcpc-8h2w-3j85
  (adm-zip, transitive of onnxruntime-node whose range cannot resolve
  the fix; same version on main), mirroring the otel entry pattern
