+++ .changeset/ontology-workbench-split.md
@@ -1,0 +1,6 @@
+---
+"@beep/ontology-ui": patch
+"@beep/ontology-client": patch
+---
+
+Split the ontology workbench into reusable region components and relocate its shared form atoms to the client package.
