+++ goals/INDEX.md
@@ -29,7 +29,7 @@ drifts from the manifests, and merge conflicts are resolved by regenerating.
 | [law-docketing-reliability](./law-docketing-reliability/README.md) | Law Docketing Reliability | 0/5 | 2026-07-14 | Prove years-out reminders cannot fail silently through escalation, acknowledgment, bounded recovery, and an independent… |
 | [law-time-capture-spine](./law-time-capture-spine/README.md) | Law Time Capture Spine | 0/5 | 2026-07-14 | Deliver the manual timer-to-approved-entry spine with matter/task context, evidence-bounded narrative, attorney control… |
 | [legal-document-intake](./legal-document-intake/README.md) | Legal Document Intake | 4/8 | 2026-07-14 | — |
-| [ontology-workbench-migration](./ontology-workbench-migration/README.md) | Ontology Workbench Migration | 1/5 | 2026-07-17 | Migrate the Ontology workbench monolith into nine fine-grained dock panels (workspace-substrate step 4, first surface) … |
+| [ontology-workbench-migration](./ontology-workbench-migration/README.md) | Ontology Workbench Migration | 3/5 | 2026-07-17 | Migrate the Ontology workbench monolith into nine fine-grained dock panels (workspace-substrate step 4, first surface) … |
 | [professional-desktop-adversarial-qa](./professional-desktop-adversarial-qa/README.md) | Professional Desktop Adversarial QA | 1/5 | 2026-07-12 | Run adversarial QA and repair loops over the professional desktop surface until two consecutive rounds are clean. |
 | [projection-dispatch-core](./projection-dispatch-core/README.md) | Projection Dispatch Core | 0/4 | 2026-07-14 | Deliver one restart-safe accepted-record projection cycle with durable target isolation and a disposable authenticated … |
 | [secure-document-delivery](./secure-document-delivery/README.md) | Secure Document Delivery | 0/4 | 2026-07-14 | Deliver an incubated secure-document-delivery capability and authenticated desktop route, proven by one authorized fixt… |
