+++ packages/ontology/ui/src/aggregates/Session/index.ts
@@ -6,16 +6,66 @@
  */
 
 /**
- * Ontology Session tree view-model exports.
- *
  * @category components
  * @since 0.0.0
  */
-export * from "./Session.tree.js";
+export { OntologyChangeLogRegion } from "./Session.changelog.js";
 /**
- * Ontology workbench screen exports.
- *
  * @category components
  * @since 0.0.0
  */
-export * from "./Session.workbench.js";
+export { OntologyDocumentRegion } from "./Session.document.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyExplorerRegion } from "./Session.explorer.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyGraphRegion } from "./Session.graph.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyInspectorRegion } from "./Session.inspector.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyMetricsRegion } from "./Session.metrics.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologySourceRegion } from "./Session.source.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologySparqlRegion } from "./Session.sparql.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { ontologyTreeItemsFor } from "./Session.tree.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyValidationRegion } from "./Session.validation.js";
+/**
+ * @category components
+ * @since 0.0.0
+ */
+export { OntologyWorkbench } from "./Session.workbench.js";
+/**
+ * @category predicates
+ * @since 0.0.0
+ */
+/**
+ * @category forms
+ * @since 0.0.0
+ */
+export { iriFieldValid, valueFromEvent } from "./Session.workbench.shared.js";
