+++ packages/ontology/ui/src/index.ts
@@ -22,9 +22,22 @@
 export const VERSION = "0.0.0" as const;
 
 /**
- * Ontology workbench UI components.
+ * Curated ontology workbench exports.
  *
  * @category components
  * @since 0.0.0
  */
-export * from "./aggregates/Session/index.js";
+export {
+  iriFieldValid,
+  OntologyChangeLogRegion,
+  OntologyDocumentRegion,
+  OntologyExplorerRegion,
+  OntologyGraphRegion,
+  OntologyInspectorRegion,
+  OntologyMetricsRegion,
+  OntologySourceRegion,
+  OntologySparqlRegion,
+  OntologyValidationRegion,
+  OntologyWorkbench,
+  ontologyTreeItemsFor,
+} from "./aggregates/Session/index.js";
